1.2: merged r10006
git-svn-id: file:///var/www/production/sfweb/svn/branches/1.2@10008 ee427ae8-e902-0410-961c-c3ed070cd9f9

+++ test/bin/prove.php
@@ -25,9 +25,12 @@ class lime_symfony extends lime_harness
 }
 
 require_once(dirname(__FILE__).'/../../lib/util/sfToolkit.class.php');
-foreach (glob(sfToolkit::getTmpDir().DIRECTORY_SEPARATOR.'/sf_autoload_unit_*') as $file)
+if($files = glob(sfToolkit::getTmpDir().DIRECTORY_SEPARATOR.'/sf_autoload_unit_*'))
 {
-  unlink($file);
+  foreach ($files as $file)
+  {
+    unlink($file);
+  }
 }
 
 // update sfCoreAutoload
